+++ aioamqp/tests/test_connection_lost.py
@@ -25,8 +25,7 @@ class ConnectionLostTestCase(testcase.RabbitTestCase, unittest.TestCase):
         self.assertTrue(amqp.is_open)
         self.assertTrue(channel.is_open)
         transport.close()  # this should have the same effect as the tcp connection being lost
-        yield from asyncio.sleep(1, loop=self.loop)
+        yield from asyncio.wait_for(amqp.worker, 1, loop=self.loop)
         self.assertFalse(amqp.is_open)
         self.assertFalse(channel.is_open)
-        yield from amqp.close()
         self.assertTrue(self.callback_called)